+++ cloud/CHANGES.md
@@ -4,6 +4,7 @@
 
 - Add the ability to download specific object versions using a `Downloader`.
 - Add more descriptive error message to the `TestRequireKeyNotFound` test helper.
+- Fix `TestClient.IterateObject` directory grouping
 
 ## v7.2.0